refactor: add all utils to index
